Abstract out some common methods used by preferences handlers.
